@@ -1,26 +1,48 @@
 {
   "name": "substance-texture",
+  "productName": "Texture",
   "version": "1.0.0-alpha.4-preview.24",
   "description": "A scientific word processor.",
-  "main": "dist/texture.cjs.js",
+  "main": "./dist/main.js",
   "jsnext:main": "index.es.js",
   "dependencies": {
     "lodash": "^4.14.1",
-    "substance": "1.0.0-preview.32"
+    "substance": "1.0.0-preview.32",
+    "dar-server": "^0.4.1"
   },
   "devDependencies": {
-    "dar-server": "^0.4.1",
-    "electron": "^1.7.9",
+    "npm-run-all": "^3.1.0",
+    "electron": "2.0.0-beta.2",
+    "electron-builder": "20.4.0",
+    "source-map-support": "0.5.3",
     "eslint": "^3.17.1",
     "font-awesome": "4.5.0",
     "substance-bundler": "^0.20.10",
     "substance-test": "^0.9.4"
   },
+  "build": {
+    "appId": "com.texture.texture-desktop",
+    "files": [
+      "dist/**/*",
+      "node_modules/**/*",
+      "package.json"
+    ],
+    "asar": "false",
+    "directories": {
+      "output": "app-dist",
+      "app": "."
+    },
+    "publish": null
+  },
   "scripts": {
+    "postinstall": "install-app-deps",
     "prepublish": "node make publish",
-    "bundle": "node make",
+    "build": "node make",
+    "watch": "node make -w",
     "start": "node make -s -w -d data",
-    "app": "node make build:app && node ./app/cli.js",
+    "release": "node make && build",
+    "electron": "electron .",
+    "app": "node make && electron .",
     "test": "npm run lint && node make test",
     "lint": "eslint -c .eslintrc.js src index.es.js"
   },
